+++ srcs/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: abdait-m <abdait-m@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/27 18:18:02 by majermou          #+#    #+#             */
-/*   Updated: 2021/03/13 17:10:53 by abdait-m         ###   ########.fr       */
+/*   Updated: 2021/03/20 14:01:24 by abdait-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 #define BUFFER_SIZE 100
 int			get_next_line(int fd, char **line);
 int			check_break_line(char *tab);
-int			ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 int			check_previous_read(char **line, char *tab);
 char		*ft_strjoin(char const *line, char const *tab);
 int			dealocate(char **line, char **tab, int ret, int var_bol);
